requirements.txt
appium-python-client>=3.0.0
pytest>=7.0.0
pytest-xdist  # для параллельного запуска тестов
